+++ lib/core/utils/helper/extention.dart
@@ -1,10 +1,7 @@
 import 'package:fitness_app/l10n/app_localizations.dart';
+import 'package:fitness_app/l10n/app_localizations_ar.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-
-extension LocalizationHelper on BuildContext {
-  AppLocalizations? get loc => AppLocalizations.of(this);
-}
 
 extension MediaQueryHelper on BuildContext {
   double get height => MediaQuery.of(this).size.height;
@@ -31,6 +28,6 @@ extension TextThemeHelper on BuildContext {
   TextTheme get textTheme => Theme.of(this).textTheme;
 }
 
-extension LocalizationHelper on BuildContext {
-  AppLocalizations get loc => AppLocalizations.of(this)!;
-}
+// extension LocalizationHelper on BuildContext {
+//   AppLocalizations get loc => AppLocalizations.of(this)!;
+// }
