+++ lib/core/utils/routes/routes_generator.dart
@@ -1,10 +1,10 @@
 import 'package:fitness_app/core/config/di.dart';
 import 'package:fitness_app/core/utils/routes/routes_name.dart';
-import 'package:fitness_app/features/home/home_view.dart';
-import 'package:fitness_app/features/layOut/lay_out.dart';
-import 'package:fitness_app/features/profile/profile_view.dart';
-import 'package:fitness_app/features/smartCoach/smart_coach_view.dart';
-import 'package:fitness_app/features/workOuts/work_outs_view.dart';
+import 'package:fitness_app/feature/home/home_view.dart';
+import 'package:fitness_app/feature/layOut/lay_out.dart';
+import 'package:fitness_app/feature/profile/profile_view.dart';
+import 'package:fitness_app/feature/smartCoach/smart_coach_view.dart';
+import 'package:fitness_app/feature/workOuts/work_outs_view.dart';
 import 'package:fitness_app/feature/login/domain/usecases/login_usecase.dart';
 import 'package:fitness_app/feature/login/presentation/view/login_screen.dart';
 import 'package:fitness_app/feature/login/presentation/view_model/login_cubit.dart';
@@ -12,7 +12,6 @@ import 'package:fitness_app/feature/onBoarding/on_boarding_one.dart';
 import 'package:fitness_app/feature/onBoarding/on_boarding_three.dart';
 import 'package:fitness_app/feature/onBoarding/on_boarding_two.dart';
 import 'package:fitness_app/feature/splash/splash_view.dart';
-import 'package:fitness_app/onboarding.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
