+++ lib/core/utils/theme/app_colors.dart
@@ -14,8 +14,7 @@ class AppColors {
   static const Color red40 = Color(0xFFFC6B21);
   static const Color red30 = Color(0x0fffd9cc);
   static const Color red20 = Color(0x0fffece5);
-  static const Color red30 = Color(0x0fffd9cc);
-  static const Color red20 = Color(0x0fffece5);
+  
 
   // Dark Shades
   static const Color dark100 = Color(0xFF1A0000);
