+++ lib/feature/layOut/lay_out.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:fitness_app/features/home/home_view.dart';
-import 'package:fitness_app/features/profile/profile_view.dart';
-import 'package:fitness_app/features/smartCoach/smart_coach_view.dart';
-import 'package:fitness_app/features/workOuts/work_outs_view.dart';
+import 'package:fitness_app/feature/home/home_view.dart';
+import 'package:fitness_app/feature/profile/profile_view.dart';
+import 'package:fitness_app/feature/smartCoach/smart_coach_view.dart';
+import 'package:fitness_app/feature/workOuts/work_outs_view.dart';
 
 class LayOut extends StatefulWidget {
   const LayOut({super.key});
