+++ lib/feature/login/presentation/view_model/login_cubit.dart
@@ -3,7 +3,6 @@ import 'package:bloc/bloc.dart';
 import 'package:flutter/material.dart';
 
 import 'package:fitness_app/core/base/api_result.dart';
-import 'package:fitness_app/core/base/base_state.dart';
 import 'package:fitness_app/feature/login/data/model/login_request.dart';
 import 'package:fitness_app/feature/login/data/model/login_response.dart';
 import 'package:fitness_app/feature/login/domain/usecases/login_usecase.dart';
@@ -23,34 +22,27 @@ class LoginCubit extends Cubit<LoginState> {
   }
 
   Future<void> _handleLogin(PerformLogin intent) async {
-    
     if (!(formKey.currentState?.validate() ?? true)) return;
-      emit(LoginLoading());
-      final response = await loginUseCase.invoke(
-          loginRequest: intent.request);
-      switch (response) {
-        case ApiError<LoginResponse>():
-          emit(
-            LoginFailure(
-              response.failure?.errorMessage ?? 'An error occurred',
-            ),
-          );
-        case ApiSuccess<LoginResponse>():
-          emit(LoginSuccess(response.data!));
-        default:
-          emit(const LoginFailure('Unexpected error occurred'));
-          break;
-      }
-    
+    emit(LoginLoading());
+    final response = await loginUseCase.invoke(loginRequest: intent.request);
+    switch (response) {
+      case ApiError<LoginResponse>():
+        emit(
+          LoginFailure(response.failure?.errorMessage ?? 'An error occurred'),
+        );
+      case ApiSuccess<LoginResponse>():
+        emit(LoginSuccess(response.data!));
+      default:
+        emit(const LoginFailure('Unexpected error occurred'));
+        break;
+    }
   }
-  }
+}
 
 sealed class LoginIntent {}
 
 class PerformLogin extends LoginIntent {
-LoginRequest request;
+  LoginRequest request;
 
-  PerformLogin({
-    required this.request,
-  });
+  PerformLogin({required this.request});
 }
