+++ lib/feature/login/presentation/view_model/login_state.dart
@@ -1,8 +1,6 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:equatable/equatable.dart';
-import 'package:fitness_app/core/base/base_state.dart';
 import 'package:fitness_app/feature/login/data/model/login_response.dart';
-
 
 sealed class LoginState extends Equatable {
   const LoginState();
@@ -12,13 +10,16 @@ sealed class LoginState extends Equatable {
 }
 
 final class LoginInitial extends LoginState {}
+
 final class LoginLoading extends LoginState {}
+
 final class LoginSuccess extends LoginState {
   final LoginResponse response;
   const LoginSuccess(this.response);
   @override
   List<Object> get props => [response];
 }
+
 final class LoginFailure extends LoginState {
   final String error;
   const LoginFailure(this.error);
