+++ lib/feature/smartCoach/smart_coach_view.dart
